Pass dependencies of analyzers as analyzer DLLs
If an analyzer library has dependencies, then those should be passed as
analyzers to the compiler. Note that this has to happen during assembly
creation and not when another library uses this analyzer, because we
need to use the actual runtime libraries present in
DotnetAssemblyRuntimeInfo as analyzers (not the refs, which is all
that's present in DotnetAssemblyCompileInfo).

In this change, we only do one level of discovery. This means if a
transitive dependency of another library is required in an analyzer
library, those must also be passed into deps. This is similar to
strict_deps, but is stricter because unlike a regular dependency which
only needs its reference DLL to be present at compile time (and
reference DLLs don't require their own dependencies), we need the full
transitive dependency chain to be present at compile time for an
analyzer.