Map embedded resource names like MSBuild
MSBuild maps the file names as "{project name}.{dotted path}" where
the project name is the same as the final generated assembly name and
the dotted path is the project-relative path to the embedded resource
with slashes changed to periods. Without this, the resource only gets
embedded with its basename in the manifest which breaks expectations
compared to MSBuild.