publish_binary: Avoid overwriting files
I hit this issue when published a binary that included a newer version
of an assembly present in the targeting pack (System.IO.Pipelines) as
a dependency. The user-provided dependency would be copied, and then
the targeting-pack provided version would overwrite it silently. This
commit addresses this issue in two ways:

1. Avoid adding the file to the script if it was previously copied.
2. As a failsafe, refuse to overwrite files if they already exist.