dotnet_tool: Add new rule
This is the low-level rule API that allows running NuGet tools using the
hermetic dotnet toolchain.

A somewhat-unusual aspect here is that instead of selecting the target
framework with select(), we instead forward dictionaries keyed by the
TFM to the rule. This is because there doesn't appear to be a way to
transition the selected framework based on the toolchain in use, so we
need to inspect the toolchain's runtime_tfm information to determine
which version of the tool we should use.